Fix handling of the mqtt client's payload buffers
The `char* payload` buffer given to the onMessage callback is the raw
TCP buffer, and not a null-terminated string as I'd expected. While we
can technically receive payloads larger than the tcp window, we know
that all valid payloads will be 6 characters or less so we don't bother
checking `index` and `total`, and just blindly use the `len`.

The `upcase` function is now pretty specific to this situation, so it'll
probably warrant some refactoring later.